Cast scale to float32 in Triton dequant kernel for numerical consistency
Addresses PR #30 review: explicitly cast scale to float32 before
multiplication, matching torch and cuTile backends.

+++ benchmarks/operators/weight_dequant/impl_triton.py
@@ -21,7 +21,7 @@ def dequant_kernel(X, S, Y, M: tl.constexpr, N: tl.constexpr,
     s_col = col // TILE_SIZE
 
     x = tl.load(X + offsets, mask=mask).to(tl.float32)
-    scale = tl.load(S + s_row * S_COLS + s_col, mask=mask)
+    scale = tl.load(S + s_row * S_COLS + s_col, mask=mask).to(tl.float32)
 
     y = x * scale
